Fixes. Relocate SUITable.
