+++ common/src/main/scala/reactST/reactTable/SUITable.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2016-2021 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package explore.config
+package reactST.reactTable
 
 import cats.syntax.all._
 import japgolly.scalajs.react._
