+++ model-testkit/shared/src/main/scala/explore/model/arb/package.scala
@@ -13,7 +13,6 @@ object all
     with ArbKeyedIndexedTree
     with ArbUserVault
     with ArbObsSummary
-    // with ArbConstraintSetModel
     with ArbConstraintsSummary
     with ArbElevationRange
     with ArbPointingId
